Move MEF exports and abstract methods to the actual implementation, and flatten
